melhorias extras: - move requisição de fetch para SWR. - reconfigura ordenação no ESLint.
