+++ .eslintrc.js
@@ -12,33 +12,21 @@ module.exports = {
           'type',
           'react',
           'react-libs',
-          ['builtin', 'external'],
           'api',
           'models',
           'shared',
           'components',
           'screens',
-          'internal-type',
           'internal',
-          ['parent-type', 'sibling-type', 'index-type'],
-          ['parent', 'sibling', 'index'],
-          'side-effect',
-          'style',
-          'object',
-          'unknown',
         ],
         'custom-groups': {
           value: {
             react: ['react', 'react-native'],
-            'react-libs': ['react-*'],
             api: ['@api/**'],
             models: ['@models/**'],
             shared: ['@shared/**'],
             components: ['@components/**'],
             screens: ['@screens/**'],
-          },
-          type: {
-            react: 'react',
           },
         },
         'newlines-between': 'always',
