+++ package.json
@@ -28,7 +28,8 @@
     "react-native-reanimated": "^3.5.4",
     "react-native-safe-area-context": "^4.7.2",
     "react-native-screens": "^3.25.0",
-    "styled-components": "^6.1.0"
+    "styled-components": "^6.1.0",
+    "swr": "^2.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
