+++ src/api/CreditCardApi.ts
@@ -1,8 +1,21 @@
-import axios, {AxiosResponse} from 'axios';
-
 import CreditCardModel from '@models/CreditCardModel';
 
+import {setSelectedCreditCard} from '@shared/Storage';
+
+import axios, {AxiosResponse} from 'axios';
+import {Fetcher} from 'swr';
+
+export const API_URL = 'http://192.168.31.159:3000/cards';
+
+export const fetcher: Fetcher<CreditCardModel[]> = async (endpoint: string) => {
+  const response = await fetch(endpoint);
+  const data = await response.json();
+  setSelectedCreditCard(data[0]);
+  return data;
+};
+
 const API_BASE_URL: string = 'http://192.168.31.159:3000';
+
 export const getCards = async (): Promise<CreditCardModel[]> => {
   try {
     const response: AxiosResponse<CreditCardModel[]> = await axios.get(
