+++ src/components/FormControlInput/index.tsx
@@ -1,3 +1,10 @@
+import {COLORS} from '@shared/Defaults';
+import globalStyles from '@shared/GlobalStyles';
+
+import {ErrorLabel, Label} from '@components/StyledComponents';
+
+import {AutoCapitalizeOptions} from 'react-native-mask-text/lib/typescript/src/@types/AutoCapitalizeOptions';
+
 import React, {FC} from 'react';
 import {KeyboardTypeOptions, View} from 'react-native';
 import {TextInput, TextInputProps} from 'react-native';
@@ -10,13 +17,6 @@ import {
   ValidationRule,
 } from 'react-hook-form';
 import {MaskedTextInput, MaskedTextInputProps} from 'react-native-mask-text';
-
-import {AutoCapitalizeOptions} from 'react-native-mask-text/lib/typescript/src/@types/AutoCapitalizeOptions';
-
-import {COLORS} from '@shared/Defaults';
-import globalStyles from '@shared/GlobalStyles';
-
-import {ErrorLabel, Label} from '@components/StyledComponents/StyledComponents';
 
 interface Props<T extends FieldValues> {
   autoCapitalize?: AutoCapitalizeOptions;
