+++ src/screens/CreditCardList/index.tsx
@@ -1,6 +1,5 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useRef, useState} from 'react';
 import {
-  Alert,
   Animated,
   FlatList,
   Text,
@@ -8,37 +7,28 @@ import {
   View,
 } from 'react-native';
 
-import styles from './styles';
-
-import {getCards} from '@api/CreditCardApi';
+import {API_URL, fetcher} from '@api/CreditCardApi';
 
 import CreditCardModel from '@models/CreditCardModel';
 
-import {COLORS, ERROR_MESSAGES} from '@shared/Defaults';
+import {COLORS} from '@shared/Defaults';
 
 import AnimatedCreditCard from '@components/AnimatedCreditCard';
 import CreditCard from '@components/CreditCard';
 import CustomButton from '@components/CustomButton';
-import {
-  BottomHeader,
-  BottomHeaderText,
-} from '@components/StyledComponents/StyledComponents';
+import {BottomHeader, BottomHeaderText} from '@components/StyledComponents';
 
-// @ts-ignore
-function CreditCardList({navigation}) {
-  const [creditCards, setCreditCards] = useState<CreditCardModel[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [selectedCreditCard, setSelectedCreditCard] = useState<CreditCardModel>(
-    {} as CreditCardModel,
-  );
+import styles from './styles';
+import useSWR from 'swr';
+
+function CreditCardList() {
+  const [selectedCreditCard, setSelectedCreditCard] = useState<
+    CreditCardModel | undefined
+  >(undefined);
   const y = new Animated.Value(0);
   const onScroll = Animated.event([{nativeEvent: {contentOffset: {y}}}], {
     useNativeDriver: true,
   });
-
-  useEffect(() => {
-    fetchData();
-  }, []);
 
   const flatListRef = useRef<FlatList | null>(null);
   const scrollToTop = () => {
@@ -47,24 +37,7 @@ function CreditCardList({navigation}) {
     }
   };
 
-  const fetchData = async () => {
-    try {
-      setLoading(true);
-      const data: CreditCardModel[] = await getCards();
-      setCreditCards(data);
-      setSelectedCreditCard(data[0]);
-      setLoading(false);
-    } catch (e) {
-      Alert.alert(ERROR_MESSAGES.API.title, ERROR_MESSAGES.API.message, [
-        {
-          text: 'OK',
-          onPress: () => {
-            navigation.navigate('Home');
-          },
-        },
-      ]);
-    }
-  };
+  const {data, isLoading} = useSWR<CreditCardModel[], Error>(API_URL, fetcher);
 
   return (
     <View style={styles.generalContainer}>
@@ -72,8 +45,9 @@ function CreditCardList({navigation}) {
         <BottomHeaderText>Meus Cartões</BottomHeaderText>
       </BottomHeader>
 
-      <View style={styles.selectedCreditCardContainer}>
-        {selectedCreditCard && !loading && (
+      <View
+        style={selectedCreditCard ? styles.selectedCreditCardContainer : []}>
+        {selectedCreditCard && (
           <>
             <CreditCard creditCard={selectedCreditCard} />
             <CustomButton
@@ -86,13 +60,13 @@ function CreditCardList({navigation}) {
       </View>
 
       <View style={styles.cardListContainer}>
-        {loading ? (
+        {isLoading ? (
           <Text>Carregando... </Text>
         ) : (
           <Animated.FlatList
             ref={flatListRef}
             scrollEventThrottle={16}
-            data={creditCards}
+            data={data}
             renderItem={({index, item}) => (
               <TouchableWithoutFeedback
                 onPress={() => {
