+++ src/screens/index.ts
@@ -2,4 +2,3 @@ export {default as AddCard} from './NewCreditCard';
 export {default as CreditCardList} from './CreditCardList';
 export {default as Home} from './Home';
 export {default as NewCreditCardForm} from './NewCreditCardForm';
-export {default as SelectedCard} from './SelectedCard';
